cw1/requirements.txt
Django>=5.2.12
 djangorestframework>=3.14
